\name{DispatchOptimizationEngine-package}
\alias{DispatchOptimizationEngine-package}
\alias{DispatchOptimizationEngine}
\docType{package}
\title{
\packageTitle{DispatchOptimizationEngine}
}
\description{
\packageDescription{DispatchOptimizationEngine}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DispatchOptimizationEngine}
\packageIndices{DispatchOptimizationEngine}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DispatchOptimizationEngine}

Maintainer: \packageMaintainer{DispatchOptimizationEngine}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ 
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
package 
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
